third_party/blink/web_tests/external/wpt/html/syntax/parsing/html5lib_template-expected.txt

This is a testharness.js-based test.
FAIL html5lib_template.html 010950d55f4eccf16e9c4af1d263bb747294c646 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>\n|       content\n|         \"Hello\"" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>"
FAIL html5lib_template.html a838bd54410cef059a42eea9606356488e16535b assert_equals: expected "#document\n| <html>\n|   <head>\n|     <template>\n|       content\n|         \"Hello\"\n|   <body>" but got "#document\n| <html>\n|   <head>\n|     <template>\n|   <body>"
FAIL html5lib_template.html 27fb9111f6675a7e033b867480c0afddcda161a6 assert_equals: expected "#document\n| <html>\n|   <head>\n|     <template>\n|       content\n|   <body>\n|     <div>" but got "#document\n| <html>\n|   <head>\n|     <template>\n|   <body>\n|     <div>"
FAIL html5lib_template.html aee883a65775489399a003b2371d58248a6aff6f assert_equals: expected "#document\n| <html>\n|   <head>\n|     <template>\n|       content\n|         \"Hello\"\n|   <body>" but got "#document\n| <html>\n|   <head>\n|     <template>\n|   <body>"
FAIL html5lib_template.html 89b17b54ab343191bf74ef5434f4d2cfac40ea97 assert_equals: expected "#document\n| <html>\n|   <head>\n|     <template>\n|       content\n|         <div>\n|   <body>" but got "#document\n| <html>\n|   <head>\n|     <template>\n|   <body>"
FAIL html5lib_template.html c4433556c7414cfd71f27b420f1ffc4348774f5e assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <div>\n|       <template>\n|         content\n|           <div>\n|             <span>\n|       <b>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <div>\n|       <template>\n|       <b>"
FAIL html5lib_template.html 3dcce7d97108b3e9ea7fa96f240ac62bf280e74b assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <div>\n|       <template>\n|         content\n|           \"Hello\"" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <div>\n|       <template>"
PASS html5lib_template.html a1f587f7ea85ccfe294bd45bfb501e850cb979e0
FAIL html5lib_template.html cd26a7832f13bdc135697321ca6c2fecdca6ef5d assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <template>\n|         content" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <template>"
FAIL html5lib_template.html e30571d90b0e56864499961eb7be955994cf72e2 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <template>\n|         content" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <template>"
FAIL html5lib_template.html 01cbe9f6a25f286b08d8dc4f7b65421e8eb3500c assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <div>\n|       <template>\n|         content\n|     <table>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <div>\n|       <template>\n|     <table>"
FAIL html5lib_template.html 96cbbcdffe02c86a8b929604c2fd5f3571a18dbe assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <div>\n|     <table>\n|       <template>\n|         content" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <div>\n|     <table>\n|       <template>"
FAIL html5lib_template.html d51676f55550e960dd0f5fa7fd0bdfa20bdde046 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       \"   \"\n|       <template>\n|         content" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       \"   \"\n|       <template>"
FAIL html5lib_template.html f9dfd9acfd494489c899604649a01d864741f50f assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <tbody>\n|         <template>\n|           content" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <tbody>\n|         <template>"
FAIL html5lib_template.html ea00361c265d3ffb47ce636d919c94ca10d58911 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <tbody>\n|         <template>\n|           content" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <tbody>\n|         <template>"
FAIL html5lib_template.html d8ebfcf7694c9d04457e796ac73049210313602e assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <tbody>\n|         <template>\n|           content" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <tbody>\n|         <template>"
FAIL html5lib_template.html b4d5e6fe9b92e2c8f54199d7cab3da383c42add0 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <thead>\n|         <template>\n|           content" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <thead>\n|         <template>"
FAIL html5lib_template.html 07724ef8f7a4fa61c77ffcd5180d3101c4781502 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <tfoot>\n|         <template>\n|           content" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <tfoot>\n|         <template>"
FAIL html5lib_template.html e90f8aae8fc690540b42b3ffa3e741e7c1dfbf43 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <select>\n|       <template>\n|         content" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <select>\n|       <template>"
FAIL html5lib_template.html 687bdf4adda88a316ec69fe20e84720acc5d1fe6 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <select>\n|       <template>\n|         content\n|           <option>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <select>\n|       <template>"
FAIL html5lib_template.html 5b232642f472c2b4c0c7511fed464eebe686b427 assert_equals: expected "#document\n| <html>\n|   <head>\n|     <template>\n|       content\n|         <option>\n|         <option>\n|   <body>" but got "#document\n| <html>\n|   <head>\n|     <template>\n|   <body>"
FAIL html5lib_template.html dc1ac1830a881d1532a1e6fd6d0cfa56d6571da2 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <select>\n|       <template>\n|         content\n|       <option>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <select>\n|       <template>\n|       <option>"
FAIL html5lib_template.html c58747a85e8b4f44d7ae63c04cdad783a903c25e assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <select>\n|       <option>\n|         <template>\n|           content" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <select>\n|       <option>\n|         <template>"
FAIL html5lib_template.html ca59bfdaec7451f704973176fab46e582bd691b2 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <select>\n|       <template>\n|         content" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <select>\n|       <template>"
FAIL html5lib_template.html cf807d6391a58c172b6c15c3b01d2a99ec0e6cf8 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <select>\n|       <option>\n|       <template>\n|         content" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <select>\n|       <option>\n|       <template>"
FAIL html5lib_template.html 350b7ac850e46de79615308fc923649264406104 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <select>\n|       <option>\n|       <template>\n|         content\n|           <option>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <select>\n|       <option>\n|       <template>"
FAIL html5lib_template.html a31ff44edf7f377543dabdda8141cda9bb6de134 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <thead>\n|         <template>\n|           content\n|             <td>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <thead>\n|         <template>"
FAIL html5lib_template.html 533c5c1b5f0d0cbb1ede2cc5ae927095c5b21f0b assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <template>\n|         content\n|           <thead>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <template>"
FAIL html5lib_template.html 61f79e083005007853c4f8e431559ac8d3845cfd assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <template>\n|         content\n|           <td>\n|             <div>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <template>"
FAIL html5lib_template.html e802e85f36792b176b73c102c0e8761d9478621d assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <template>\n|         content\n|           <thead>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <template>"
FAIL html5lib_template.html 51d0797ff7653cd7be34458d689146e08a666c7f assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <thead>\n|         <template>\n|           content\n|             <tr>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <thead>\n|         <template>"
FAIL html5lib_template.html d60e4079a18bd6266740cc61d1ca736e9d5098ce assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <template>\n|         content\n|           <tr>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <template>"
FAIL html5lib_template.html 308709292677b4d74c108a811ad7b7acd0bdfc9c assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <tbody>\n|         <tr>\n|           <template>\n|             content\n|               <td>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <tbody>\n|         <tr>\n|           <template>"
FAIL html5lib_template.html 8965cdf9c4e9936262e25c90c7a7f8673840a445 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <template>\n|         content\n|           <tr>\n|             <template>\n|               content\n|                 <td>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <template>"
FAIL html5lib_template.html 7dccda789764beb489e09be10188af9347335d05 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <template>\n|         content\n|           <tr>\n|             <template>\n|               content\n|                 <td>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <template>"
FAIL html5lib_template.html e15be51c77e1a6de35568a099ed339440ce9426d assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <template>\n|         content\n|           <td>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <template>"
FAIL html5lib_template.html 503d3782e45940c19f096f360a092282b46ab1ea assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>\n|       content\n|         <td>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>"
FAIL html5lib_template.html b4ab56fd9e9cebf479d14adfa523c06d16483a5e assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>\n|       content\n|         <template>\n|           content\n|             <tr>\n|         <td>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>"
FAIL html5lib_template.html cd8bc9521f9683086a9e8529dd97314a6869daeb assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <colgroup>\n|         <template>\n|           content\n|             <col>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <colgroup>\n|         <template>"
FAIL html5lib_template.html f915e7b3407c24b28c3aad318e5693cc774020f4 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <frameset>\n|     <frame>" but got "#document\n| <html>\n|   <head>\n|   <frameset>\n|     <template>"
FAIL html5lib_template.html 3c5eb261787b3d15aff86fa61de773fd7e439b0e assert_equals: expected "#document\n| <html>\n|   <head>\n|     <template>\n|       content\n|   <body>" but got "#document\n| <html>\n|   <head>\n|     <template>\n|   <body>"
FAIL html5lib_template.html 2b57775750c198d4b98b23aed74ff80a866a01f5 assert_equals: expected "#document\n| <html>\n|   <head>\n|     <template>\n|       content\n|         <div>\n|           <span>\n|         <span>\n|   <body>" but got "#document\n| <html>\n|   <head>\n|     <template>\n|   <body>"
FAIL html5lib_template.html dc3d016610f3ab532525a6c2871f03d6b62b0168 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>\n|       content\n|         <div>\n|           <span>\n|         <span>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>"
FAIL html5lib_template.html 6a184d71d00580a26a8b6bd97aafe5503339f3f6 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>\n|       content\n|         <script>\n|           \"var i = 1;\"\n|         <td>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>"
FAIL html5lib_template.html ce570a6c4bcee8b72a03e25508c6dd72e3cc6c35 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>\n|       content\n|         <tr>\n|         <div>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>"
FAIL html5lib_template.html e0c3d922f7b1f1654f02f716c3d9b31198ce3385 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>\n|       content\n|         <tr>\n|         <tr>\n|           <td>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>"
FAIL html5lib_template.html 87e67242bf6debcf3b7dca852d10aa0f7b625b28 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>\n|       content\n|         <td>\n|         <td>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>"
FAIL html5lib_template.html 35ac4d4c972a01d368ed0cacb41370efef0a644d assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>\n|       content\n|         <td>\n|         <td>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>"
FAIL html5lib_template.html 5226c39dfc2d624ad4191b4eacb7e40c7ae528eb assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>\n|       content\n|         <td>\n|         <td>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>"
FAIL html5lib_template.html aa90cd4db6b12e0a47341914a90cc536eec32d64 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>\n|       content\n|         <td>\n|         <td>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>"
FAIL html5lib_template.html 48af1faf5fcf48a0854af5a5c33656d9ccf6736b assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>\n|       content\n|         <td>\n|         <td>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>"
FAIL html5lib_template.html ed3a029ba5e7f59969d65a4fc490a8f13b098cb9 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>\n|       content\n|         <tr>\n|         <tr>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>"
FAIL html5lib_template.html 6c8880d54475ad9574e203dcf2e55820b123cc64 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>\n|       content\n|         <tr>\n|         <tr>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>"
FAIL html5lib_template.html 275060925a844cb51b29bae660301de9780d68c8 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>\n|       content\n|         <tr>\n|         <tr>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>"
FAIL html5lib_template.html 9f82f6ec4c0a48c1d4dfbe6803b94abd553aea88 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>\n|       content\n|         <thead>\n|         <caption>\n|         <tbody>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>"
FAIL html5lib_template.html f094bf7e94a88b86c80a0643e70c8e5ff3354698 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>\n|       content\n|         <thead>\n|         <tbody>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>"
FAIL html5lib_template.html 35a07ec3b4bf26ea407dc1ddf52f14195a714059 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>\n|       content\n|         <div>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>"
FAIL html5lib_template.html 24faa53b271f994a4ff31d5796c8ff47d6f2c3e6 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>\n|       content\n|         <em>\n|           \"Hello\"" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>"
FAIL html5lib_template.html 0f1c491b58c2dd3c402a62e37f833bc1f1db8d21 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>\n|       content\n|         <!-- comment -->" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>"
FAIL html5lib_template.html 868d918a7b5d8b5c065c15229492bc2022bfbcba assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>\n|       content\n|         <style>\n|         <td>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>"
FAIL html5lib_template.html 0538efa44e857596c556033a3821d424378aea3f assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>\n|       content\n|         <meta>\n|         <td>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>"
FAIL html5lib_template.html e7d7bf3973c70d3cf9b0adad2ebed9f25be48d66 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>\n|       content\n|         <link>\n|         <td>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>"
FAIL html5lib_template.html c69d0ac542d477b7312bb24981127b8aa8fdb1df assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <colgroup>\n|         <template>\n|           content\n|             <col>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <colgroup>\n|         <template>"
FAIL html5lib_template.html b496a8c13a7bd75b778bb0de489726aee952ae0c assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     a=\"b\"\n|     <template>\n|       content\n|         <div>\n|         <div>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     a=\"b\"\n|     <template>"
FAIL html5lib_template.html 5d6ee61de40274c9626ca78ee208d51276d5662d assert_equals: expected "#document\n| <html>\n|   a=\"b\"\n|   <head>\n|     <template>\n|       content\n|         <div>\n|           <span>\n|   <body>" but got "#document\n| <html>\n|   a=\"b\"\n|   <head>\n|     <template>\n|   <body>"
FAIL html5lib_template.html 9bd9687a65f258adc24450fc5cbd781fff6c038a assert_equals: expected "#document\n| <html>\n|   a=\"b\"\n|   <head>\n|     <template>\n|       content\n|         <col>\n|         <col>\n|   <body>" but got "#document\n| <html>\n|   a=\"b\"\n|   <head>\n|     <template>\n|   <body>"
FAIL html5lib_template.html db1baeb846d718c773324746524fbd68f2e9436e assert_equals: expected "#document\n| <html>\n|   a=\"b\"\n|   <head>\n|     <template>\n|       content\n|   <body>" but got "#document\n| <html>\n|   a=\"b\"\n|   <head>\n|     <template>\n|   <body>"
FAIL html5lib_template.html 4b0ce46c611dbcc016db272ef007f302bee0c897 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>\n|       content\n|         <tr>\n|         <template>\n|           content\n|         <tr>\n|           <td>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>"
FAIL html5lib_template.html 1a735e1c7f28f8701f3c7fd5e9404b8911916086 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>\n|       content\n|         <thead>\n|         <template>\n|           content\n|             <tr>\n|         <tbody>\n|           <tr>\n|         <tfoot>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>"
FAIL html5lib_template.html 0686eedec06b2db1dc283fac92c1ef1a33114c71 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>\n|       content\n|         <template>\n|           content\n|             <b>\n|               <template>\n|                 content\n|         \"text\"" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>"
FAIL html5lib_template.html d4dfb87ce626f12923056a6cd77448eaf4660ac2 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>\n|       content\n|         <col>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>"
FAIL html5lib_template.html 1f295920f2937b2c8023b3761c43a0d4d9e5353c assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>\n|       content\n|         <col>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>"
FAIL html5lib_template.html 3b91fa08fad923d387d924cff37fbf6b4c3a5712 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>\n|       content\n|         <col>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>"
FAIL html5lib_template.html 45a1c1ad5d99ad67c573096a79253996a664e01b assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>\n|       content\n|         <col>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>"
FAIL html5lib_template.html 0fe3a66773c6048c8f6f2c92f2611f65be972ec1 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>\n|       content\n|         <col>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>"
FAIL html5lib_template.html be40897ca411e1507197c31ab2a9f9752a05f769 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>\n|       content\n|         <col>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>"
FAIL html5lib_template.html dcfb1048ed5c40e406b4fbf0cde24c826713907f assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>\n|       content\n|         <i>\n|         <menu>\n|           <i>\n|             \"Foo\"" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>"
FAIL html5lib_template.html 78263aeea68ac97903598682013bae9c0c21d547 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>\n|       content\n|         <div>\n|           \"Foo\"\n|         <template>\n|           content" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>"
FAIL html5lib_template.html 5aa177ef1a35bf4502dcb867d8e666288982ba99 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <div>\n|       <template>\n|         content\n|           <tr>\n|             <td>\n|               \"Foo\"" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <div>\n|       <template>"
FAIL html5lib_template.html 5d303375907dc4d4380b477e0317c17b660613e9 assert_equals: expected "#document\n| <html>\n|   <head>\n|     <template>\n|       content\n|         <sub>\n|           <table>\n|   <body>" but got "#document\n| <html>\n|   <head>\n|     <template>\n|   <body>"
FAIL html5lib_template.html d822f726927c34b92fe102b13e63920850878f6a assert_equals: expected "#document\n| <html>\n|   <head>\n|     <template>\n|       content\n|         <template>\n|           content\n|   <body>" but got "#document\n| <html>\n|   <head>\n|     <template>\n|   <body>"
FAIL html5lib_template.html 07acdcaeb4fa639296d46673cf28823ddf2a6ca7 assert_equals: expected "#document\n| <html>\n|   <head>\n|     <template>\n|       content\n|         <div>\n|   <body>" but got "#document\n| <html>\n|   <head>\n|     <template>\n|   <body>"
FAIL html5lib_template.html 58bd846ce1be0caf7560fba2ef19e2c2070ab123 assert_equals: expected "#document\n| <html>\n|   <head>\n|     <template>\n|       content\n|         <template>\n|           content\n|             <div>\n|   <body>" but got "#document\n| <html>\n|   <head>\n|     <template>\n|   <body>"
FAIL html5lib_template.html 8eeee377e5ab324731cc592f1fa8abe1045ad610 assert_equals: expected "#document\n| <html>\n|   <head>\n|     <template>\n|       content\n|         <template>\n|           content\n|             <table>\n|   <body>" but got "#document\n| <html>\n|   <head>\n|     <template>\n|   <body>"
FAIL html5lib_template.html b30690019090149132fc228a7261c5cf2fd149fc assert_equals: expected "#document\n| <html>\n|   <head>\n|     <template>\n|       content\n|         <template>\n|           content\n|             <tbody>\n|   <body>" but got "#document\n| <html>\n|   <head>\n|     <template>\n|   <body>"
FAIL html5lib_template.html 67a209d928804f90fdb66d070201b23f3d0c8a42 assert_equals: expected "#document\n| <html>\n|   <head>\n|     <template>\n|       content\n|         <template>\n|           content\n|             <tr>\n|   <body>" but got "#document\n| <html>\n|   <head>\n|     <template>\n|   <body>"
FAIL html5lib_template.html 12104886b8f87daa937eac30b5ff0e1e074eaa6f assert_equals: expected "#document\n| <html>\n|   <head>\n|     <template>\n|       content\n|         <template>\n|           content\n|             <td>\n|   <body>" but got "#document\n| <html>\n|   <head>\n|     <template>\n|   <body>"
FAIL html5lib_template.html 483cc9957a7225fe435112642be59abb4c459a1e assert_equals: expected "#document\n| <html>\n|   <head>\n|     <template>\n|       content\n|         <template>\n|           content\n|             <caption>\n|   <body>" but got "#document\n| <html>\n|   <head>\n|     <template>\n|   <body>"
FAIL html5lib_template.html 72d8ac431a154c40ab75d53a258d9d80d47689eb assert_equals: expected "#document\n| <html>\n|   <head>\n|     <template>\n|       content\n|         <template>\n|           content\n|             <colgroup>\n|   <body>" but got "#document\n| <html>\n|   <head>\n|     <template>\n|   <body>"
FAIL html5lib_template.html 1125967cbbcd404f4cb14d48270b8ec778970d77 assert_equals: expected "#document\n| <html>\n|   <head>\n|     <template>\n|       content\n|         <template>\n|           content\n|             <col>\n|   <body>" but got "#document\n| <html>\n|   <head>\n|     <template>\n|   <body>"
FAIL html5lib_template.html 32c963e164b9ec82c60e490bb141c1ccc70b992f assert_equals: expected "#document\n| <html>\n|   <head>\n|     <template>\n|       content\n|         <template>\n|           content\n|             <tbody>\n|             <select>\n|   <body>" but got "#document\n| <html>\n|   <head>\n|     <template>\n|   <body>"
FAIL html5lib_template.html 574a95fc9c9f2de3aeaa0c9ee1e6967fc3d4770d assert_equals: expected "#document\n| <html>\n|   <head>\n|     <template>\n|       content\n|         <template>\n|           content\n|             \"Foo\"\n|             <table>\n|   <body>" but got "#document\n| <html>\n|   <head>\n|     <template>\n|   <body>"
FAIL html5lib_template.html 332863a7f9e61bff32bd3427ede7a088b790d453 assert_equals: expected "#document\n| <html>\n|   <head>\n|     <template>\n|       content\n|         <template>\n|           content\n|   <body>" but got "#document\n| <html>\n|   <head>\n|     <template>\n|   <body>"
FAIL html5lib_template.html 2121db07146781773df9e53b94fa921a805175ce assert_equals: expected "#document\n| <html>\n|   <head>\n|     <template>\n|       content\n|         <template>\n|           content\n|             <script>\n|               \"var i\"\n|   <body>" but got "#document\n| <html>\n|   <head>\n|     <template>\n|   <body>"
FAIL html5lib_template.html 8675de267cd7e34f2febdee3feb665614d1562fe assert_equals: expected "#document\n| <html>\n|   <head>\n|     <template>\n|       content\n|         <template>\n|           content\n|             <style>\n|               \"var i\"\n|   <body>" but got "#document\n| <html>\n|   <head>\n|     <template>\n|   <body>"
FAIL html5lib_template.html c5d26ad923a2b1e988ddd378ca4fb26eb48353e1 assert_equals: expected "#document\n| <html>\n|   <head>\n|     <template>\n|       content\n|         <table>\n|   <body>\n|     <span>\n|       \"Foo\"" but got "#document\n| <html>\n|   <head>\n|     <template>\n|   <body>\n|     <span>\n|       \"Foo\""
FAIL html5lib_template.html eec1542e2fa0e9eafb7f8d4a51eae56b5a31b3c8 assert_equals: expected "#document\n| <html>\n|   <head>\n|     <template>\n|       content\n|         <td>\n|   <body>\n|     <span>\n|       \"Foo\"" but got "#document\n| <html>\n|   <head>\n|     <template>\n|   <body>\n|     <span>\n|       \"Foo\""
FAIL html5lib_template.html b79387a54c3b136db0f28ed96555ff683b3947fe assert_equals: expected "#document\n| <html>\n|   <head>\n|     <template>\n|       content\n|         <object>\n|   <body>\n|     <span>\n|       \"Foo\"" but got "#document\n| <html>\n|   <head>\n|     <template>\n|   <body>\n|     <span>\n|       \"Foo\""
FAIL html5lib_template.html c477a29a4deb32d072a415fa809a84a4f2beee0c assert_equals: expected "#document\n| <html>\n|   <head>\n|     <template>\n|       content\n|         <svg svg>\n|           <svg template>\n|   <body>" but got "#document\n| <html>\n|   <head>\n|     <template>\n|   <body>"
FAIL html5lib_template.html 26e4480c08e1f5f7b6ac8b8c1832ab0312e3b7c5 assert_equals: expected "#document\n| <html>\n|   <head>\n|     <template>\n|       content\n|         <svg svg>\n|           <svg foo>\n|             <svg template>\n|               <svg foreignObject>\n|                 <div>\n|   <body>\n|     <div>" but got "#document\n| <html>\n|   <head>\n|     <template>\n|   <body>\n|     <div>"
FAIL html5lib_template.html 24b3b50fdd0bf8d5cf2ebaa6bf502d7bcfde1da4 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <dummy>\n|       <template>\n|         content\n|           <span>" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <dummy>\n|       <template>"
FAIL html5lib_template.html d3704c68528357189eb5826ab66eea071d6137a5 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <tbody>\n|         <tr>\n|           <td>\n|             <select>\n|               <template>\n|                 content\n|                   \"Foo\"\n|       <caption>\n|         \"A\"" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <table>\n|       <tbody>\n|         <tr>\n|           <td>\n|             <select>\n|               <template>\n|       <caption>\n|         \"A\""
FAIL html5lib_template.html d958f7d44faf772d1fb60f1a8f186f837ca735d9 assert_equals: expected "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>\n|       content" but got "#document\n| <html>\n|   <head>\n|   <body>\n|     <template>"
FAIL html5lib_template.html 3fc4d97fa68fc2658356bdbd4e051c867de8de53 assert_equals: expected "#document\n| <html>\n|   <head>\n|     <template>\n|       content\n|   <body>" but got "#document\n| <html>\n|   <head>\n|     <template>\n|   <body>"
FAIL html5lib_template.html 94820107bbf3fab3f82de1f717e8413aead7d3a6 assert_equals: expected "#document\n| <html>\n|   <head>\n|     <template>\n|       content\n|         \"Foo\"\n|   <body>" but got "#document\n| <html>\n|   <head>\n|     <template>\n|   <body>"
FAIL html5lib_template.html 657c00ebdda37ae060cc69633ed98482ccc29e18 assert_equals: expected "#document\n| <!DOCTYPE html>\n| <html>\n|   <head>\n|   <body>\n|     <dummy>\n|       <table>\n|         <template>\n|           content\n|             <table>\n|               <template>\n|                 content\n|                   <table>\n|                     <script>" but got "#document\n| <!DOCTYPE html>\n| <html>\n|   <head>\n|   <body>\n|     <dummy>\n|       <table>\n|         <template>"
Harness: the test ran to completion.
